// graphQL/schema.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  Dimension: { input: any; output: any; }
  HexColor: { input: any; output: any; }
  JSON: { input: any; output: any; }
  Quality: { input: any; output: any; }
};

/** Represents a binary file in a space. An asset can be any file type. */
export type Asset = {
  __typename?: 'Asset';
  contentType?: Maybe<Scalars['String']['output']>;
  contentfulMetadata: ContentfulMetadata;
  description?: Maybe<Scalars['String']['output']>;
  fileName?: Maybe<Scalars['String']['output']>;
  height?: Maybe<Scalars['Int']['output']>;
  linkedFrom?: Maybe<AssetLinkingCollections>;
  size?: Maybe<Scalars['Int']['output']>;
  sys: Sys;
  title?: Maybe<Scalars['String']['output']>;
  url?: Maybe<Scalars['String']['output']>;
  width?: Maybe<Scalars['Int']['output']>;
};
// ...
export type AssetCollection = {
  __typename?: 'AssetCollection';
  items: Array<Maybe<Asset>>;
  limit: Scalars['Int']['output'];
  skip: Scalars['Int']['output'];
  total: Scalars['Int']['output'];
};
// ...
export type AssetLinkingCollections = {
  __typename?: 'AssetLinkingCollections';
  categoryCollection?: Maybe<CategoryCollection>;
  entryCollection?: Maybe<EntryCollection>;
  galleryItemCollection?: Maybe<GalleryItemCollection>;
  imageCollection?: Maybe<ImageCollection>;
};
// ...
export type BoxContent = Entry & {
  __typename?: 'BoxContent';
  contentfulMetadata: ContentfulMetadata;
  item?: Maybe<Scalars['String']['output']>;
  linkedFrom?: Maybe<BoxContentLinkingCollections>;
  quantity?: Maybe<Scalars['Int']['output']>;
  sys: Sys;
};
// ...
export type BoxContentCollection = {
  __typename?: 'BoxContentCollection';
  items: Array<Maybe<BoxContent>>;
  limit: Scalars['Int']['output'];
  skip: Scalars['Int']['output'];
  total: Scalars['Int']['output'];
};
// ...
export type Category = Entry & {
  __typename?: 'Category';
  contentfulMetadata: ContentfulMetadata;
  linkedFrom?: Maybe<CategoryLinkingCollections>;
  name?: Maybe<Scalars['String']['output']>;
  productsCollection?: Maybe<CategoryProductsCollection>;
  slug?: Maybe<Scalars['String']['output']>;
  sys: Sys;
  thumbnail?: Maybe<Asset>;
};
// ...
export type CategoryCollection = {
  __typename?: 'CategoryCollection';
  items: Array<Maybe<Category>>;
  limit: Scalars['Int']['output'];
  skip: Scalars['Int']['output'];
  total: Scalars['Int']['output'];
};
// ...
export type CategoryLinkingCollections = {
  __typename?: 'CategoryLinkingCollections';
  entryCollection?: Maybe<EntryCollection>;
  productCollection?: Maybe<ProductCollection>;
};
// ...
export type CategoryProductsCollection = {
  __typename?: 'CategoryProductsCollection';
  items: Array<Maybe<Entry>>;
  limit: Scalars['Int']['output'];
  skip: Scalars['Int']['output'];
  total: Scalars['Int']['output'];
};

export type ContentfulMetadata = {
  __typename?: 'ContentfulMetadata';
  tags: Array<Maybe<ContentfulTag>>;
};
// ...
export type Entry = {
  contentfulMetadata: ContentfulMetadata;
  sys: Sys;
};

export type EntryCollection = {
  __typename?: 'EntryCollection';
  items: Array<Maybe<Entry>>;
  limit: Scalars['Int']['output'];
  skip: Scalars['Int']['output'];
  total: Scalars['Int']['output'];
};
// ...
export type GalleryItem = Entry & {
  __typename?: 'GalleryItem';
  contentfulMetadata: ContentfulMetadata;
  desktopCollection?: Maybe<AssetCollection>;
  linkedFrom?: Maybe<GalleryItemLinkingCollections>;
  mobileCollection?: Maybe<AssetCollection>;
  name?: Maybe<Scalars['String']['output']>;
  sys: Sys;
  tabletCollection?: Maybe<AssetCollection>;
};
// ...
export type GalleryItemCollection = {
  __typename?: 'GalleryItemCollection';
  items: Array<Maybe<GalleryItem>>;
  limit: Scalars['Int']['output'];
  skip: Scalars['Int']['output'];
  total: Scalars['Int']['output'];
};
// ...
export type Image = Entry & {
  __typename?: 'Image';
  contentfulMetadata: ContentfulMetadata;
  desktop?: Maybe<Asset>;
  linkedFrom?: Maybe<ImageLinkingCollections>;
  mobile?: Maybe<Asset>;
  name?: Maybe<Scalars['String']['output']>;
  sys: Sys;
  tablet?: Maybe<Asset>;
  tabletPdp?: Maybe<Asset>;
};
// ...
export type ImageCollection = {
  __typename?: 'ImageCollection';
  items: Array<Maybe<Image>>;
  limit: Scalars['Int']['output'];
  skip: Scalars['Int']['output'];
  total: Scalars['Int']['output'];
};
// ...
export type Product = Entry & {
  __typename?: 'Product';
  category?: Maybe<Category>;
  contentfulMetadata: ContentfulMetadata;
  description?: Maybe<ProductDescription>;
  features?: Maybe<ProductFeatures>;
  gallery?: Maybe<GalleryItem>;
  isNew?: Maybe<Scalars['Boolean']['output']>;
  linkedFrom?: Maybe<ProductLinkingCollections>;
  name?: Maybe<Scalars['String']['output']>;
  price?: Maybe<Scalars['Float']['output']>;
  productBoxContentCollection?: Maybe<ProductProductBoxContentCollection>;
  productImages?: Maybe<Image>;
  relatedProductsCollection?: Maybe<ProductRelatedProductsCollection>;
  slug?: Maybe<Scalars['String']['output']>;
  sys: Sys;
};
// ...
export type ProductCollection = {
  __typename?: 'ProductCollection';
  items: Array<Maybe<Product>>;
  limit: Scalars['Int']['output'];
  skip: Scalars['Int']['output'];
  total: Scalars['Int']['output'];
};
// ...
export type Query = {
  __typename?: 'Query';
  asset?: Maybe<Asset>;
  assetCollection?: Maybe<AssetCollection>;
  boxContent?: Maybe<BoxContent>;
  boxContentCollection?: Maybe<BoxContentCollection>;
  category?: Maybe<Category>;
  categoryCollection?: Maybe<CategoryCollection>;
  entryCollection?: Maybe<EntryCollection>;
  galleryItem?: Maybe<GalleryItem>;
  galleryItemCollection?: Maybe<GalleryItemCollection>;
  image?: Maybe<Image>;
  imageCollection?: Maybe<ImageCollection>;
  product?: Maybe<Product>;
  productCollection?: Maybe<ProductCollection>;
};
// ...
export type Sys = {
  __typename?: 'Sys';
  environmentId: Scalars['String']['output'];
  firstPublishedAt?: Maybe<Scalars['DateTime']['output']>;
  id: Scalars['String']['output'];
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  publishedVersion?: Maybe<Scalars['Int']['output']>;
  spaceId: Scalars['String']['output'];
};
// ...
export type GetCategoriesQuery = { __typename?: 'Query', categoryCollection?: { __typename?: 'CategoryCollection', items: Array<{ __typename?: 'Category', name?: string | null, thumbnail?: { __typename?: 'Asset', url?: string | null } | null } | null> } | null };